Undergraduate-Projects/ISTA498Project-master 3/project/backend/requirements.txt
asn1crypto==1.5.1
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==36.0.2
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.4
importlib-metadata==4.12.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oscrypto==1.3.0
pycparser==2.21
pycryptodomex==3.15.0
PyJWT==2.5.0
pyOpenSSL==22.0.0
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
schedule==1.1.0
six==1.16.0
snowflake-connector-python==2.8.0
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
